requirements.txt
flask
numpy
torch
Pillow
torchvision
ultralytics